+++ JustLab/Accel/OctTree.h
@@ -1,5 +1,5 @@
 //================================================
-// 层次包围体
+// 八叉树
 //================================================
 
 #pragma once
@@ -12,7 +12,7 @@ struct OctTree : public Accel {
 
   void Divide(size_t n, std::vector<AccelNode> *children) override;
   void Traverse(const Ray3f &ray) override;
-  std::pair<int, int> GetLimits() override;
+  std::pair<size_t, size_t> GetLimits() override;
 };
 
 }
